requirements.txt
aiogram~=2.15
SQLAlchemy~=1.4.25
python-dotenv~=0.19.1